requirements.txt
annotated-types==0.6.0
attrs==23.2.0
blinker==1.6.2
certifi==2024.2.2
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
frozenlist==1.4.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.5
progress==1.6
progressbar==2.5
psycopg2==2.9.6
pydantic==2.5.3
pydantic_core==2.14.6
pyTelegramBotAPI==4.12.0
requests==2.31.0
telebot==0.0.5
tqdm==4.65.0
typing_extensions==4.10.0
urllib3==2.0.3
vk-api==11.9.9
Werkzeug==2.3.6
yarl==1.9.4

aiogram~=3.4.1
SQLAlchemy~=2.0.30
redis~=5.0.4
python-dotenv~=1.0.1